Add new logging implementation
Effectively a bit off-topic for this branch, but anyway...

This implementation uses a better approach to logging, at least
in my opinion. Every logging message is on one line with all the
necessary details and even more. Also, the various messages do
not clutter anymore the "stderr" stream, while they are
differentiated in "stdout" for INFO and lower, while WARNING and
higher remain on "stderr". Also, it is still in the works a way
to output everything to its own logging file, to debug properly.

The old logging system is still available through the command
line argument --disable-varied-logging.

Also refer to the class StaticLoggerBinder for more information
about implementation details and what you should *NOT* do.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>