Refactor the software
I think that using packages is a neater way of organizing the software.
Plus, this allows more scalability when the size of the project
increases.

I know this breaks serialization, but it doesn't really matter: on
the first run of this new version, the entire cache is rebuilt from
scratch: there are no errors to speak of. Intercompatibility is not
an issue, too, because serialization is only used for caching.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>