Add IPrinter interfaces
This allows the user to specify via commands various custom printer
implementations, which is actually the main goal of this branch.

As a consequence, the end user can use printer which provide output
to the console or a special file format, everything is completely
transparent to the software.

A configurable version of a printer (IConfigurablePrinter) has been
added, so that more sophisticated implementations can request more
information to the user without having to assume anything or hack
inside the argument provider algorithm (if that's even possible).

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>