Add proprietary and temporary HTML library
This library is temporary and only used as a replacement for the
upcoming real version. Please do not rely on it for now because
it will be soon removed.

If you are wondering why a custom library has been used instead of
J2HTML is because of the way the pretty printer works. The use of
a custom library seemed more effective. Also this allows us to use
custom indentation and custom "toString" implementations to make
the entire HTML output as nice as possible. In the end, this is a
pretty printer. :+1: :joy:

Last but not least, this will allow the printer to move to the
REAL version of the library when possible and so use the full
features of the HTML environment, with strong-typing (checking the
various tags and attributes before applying them to a document)
and new version features (e.g. HTML5 video tag, if ever needed)

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>